split up kprint, other than format
