+++ game/kernel/common/fileio.h
@@ -1,2 +1,4 @@
 #pragma once
+
+char* strend(char* str);
 void kstrcpy(char* dst, const char* src);
