+++ game/kernel/common/kscheme.h
@@ -1,6 +1,10 @@
 #pragma once
 
 #include "common/common_types.h"
+
+#include "game/kernel/common/Ptr.h"
+
 extern s32 NumSymbols;
+extern Ptr<u32> s7;
 
 void kscheme_init_globals_common();
