+++ game/kernel/jak1/kprint.h
@@ -1,0 +1,12 @@
+#pragma once
+
+/*!
+ * @file kprint.h
+ * GOAL Print.  Contains GOAL I/O, Print, Format...
+ */
+
+#include "common/common_types.h"
+
+extern "C" {
+s32 format_impl(uint64_t* args);
+}
