+++ game/kernel/jak2/kprint.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void output_sql_query(char* query_name);
