+++ game/kernel/kmachine.cpp
@@ -14,7 +14,6 @@
 #include "kdgo.h"
 #include "klink.h"
 #include "klisten.h"
-#include "kprint.h"
 #include "kscheme.h"
 #include "ksound.h"
 #include "svnrev.h"
@@ -32,8 +31,10 @@
 #include "game/kernel/common/fileio.h"
 #include "game/kernel/common/kboot.h"
 #include "game/kernel/common/kdsnetm.h"
+#include "game/kernel/common/klink.h"
 #include "game/kernel/common/kmalloc.h"
 #include "game/kernel/common/kprint.h"
+#include "game/kernel/common/kscheme.h"
 #include "game/kernel/common/ksocket.h"
 #include "game/kernel/common/memory_layout.h"
 #include "game/kernel/jak1/kboot.h"
