+++ game/mips2c/functions/test_func.cpp
@@ -18,7 +18,6 @@ u64 execute(void* ctxt) {
 }  // namespace Mips2C
 
 //--------------------------MIPS2C---------------------
-#include "game/kernel/kprint.h"
 #include "game/kernel/kscheme.h"
 #include "game/mips2c/mips2c_private.h"
 namespace Mips2C {
